moved events-global global from collections dir to new gloabls dir
