+++ src/globals/events-global.ts
@@ -1,9 +1,9 @@
 import type { GlobalConfig } from 'payload'
 
 import { cacheTags } from '@/lib/utils/revalidation'
-import { anyone } from './_access/anyone'
-import { authenticated } from './_access/authenticated'
-import { customUploadField } from './_fields/custom-upload'
+import { anyone } from '../collections/_access/anyone'
+import { authenticated } from '../collections/_access/authenticated'
+import { customUploadField } from '../collections/_fields/custom-upload'
 
 export const EventsGlobal: GlobalConfig = {
   slug: 'events-global',
